+++ include/dolfin/fem/numbering/CG1s.h
@@ -26,10 +26,9 @@ class CG1sNumbering : public DofNumbering
 {
 
 public:
-
   ///
-  CG1sNumbering(Mesh& mesh, ufc::dofmap& ufc_dofmap) :
-      DofNumbering(mesh, ufc_dofmap)
+  CG1sNumbering( Mesh & mesh, ufc::dofmap & ufc_dofmap )
+    : DofNumbering( mesh, ufc_dofmap )
   {
   }
 
@@ -37,10 +36,13 @@ public:
   ~CG1sNumbering() override = default;
 
   ///
-  inline void tabulate_dofs(uint* dofs, ufc::cell const& ufc_cell, Cell const& cell) const override
+  inline void tabulate_dofs( size_t *          dofs,
+                             ufc::cell const & ufc_cell,
+                             Cell const &      cell ) const override
   {
-    std::copy(ufc_cell.entity_indices[0].data(),
-              ufc_cell.entity_indices[0].data() + cell.num_entities(0), dofs);
+    std::copy( ufc_cell.entity_indices[0].data(),
+               ufc_cell.entity_indices[0].data() + cell.num_entities( 0 ),
+               dofs );
   }
 
   ///
@@ -48,57 +50,56 @@ public:
   {
     DofNumbering::init();
     //---
-    if (ufc_dofmap.num_element_dofs() != mesh.type().num_entities(0))
+    if ( ufc_dofmap.num_element_dofs() != mesh.type().num_entities( 0 ) )
     {
-      error("CG1sNumbering : local dimension %u != %u",
-            ufc_dofmap.num_element_dofs(), mesh.type().num_entities(0));
+      error( "CG1sNumbering : local dimension %u != %u",
+             ufc_dofmap.num_element_dofs(),
+             mesh.type().num_entities( 0 ) );
     }
-    set_range(mesh.topology().offset(0), mesh.topology().num_owned(0));
+    set_range( mesh.topology().offset( 0 ), mesh.topology().num_owned( 0 ) );
     //---
-    if (mesh.is_distributed())
+    if ( mesh.is_distributed() )
     {
-      DistributedData const& distdata = mesh.distdata()[0];
-      if (!distdata.valid_numbering)
+      DistributedData const & distdata = mesh.distdata()[0];
+      if ( !distdata.valid_numbering )
       {
-        error("CG1sNumbering : vertex numbering is invalid");
+        error( "CG1sNumbering : vertex numbering is invalid" );
       }
       shared_.clear();
-      for (SharedIterator it(distdata); it.valid(); ++it)
+      for ( SharedIterator it( distdata ); it.valid(); ++it )
       {
-        shared_.insert(it.global_index());
+        shared_.insert( it.global_index() );
       }
       ghosts_.clear();
-      for (GhostIterator it(distdata); it.valid(); ++it)
+      for ( GhostIterator it( distdata ); it.valid(); ++it )
       {
-        ghosts_.insert(it.global_index());
+        ghosts_.insert( it.global_index() );
       }
     }
   }
 
   ///
-  inline bool is_shared(uint index) const override
+  inline bool is_shared( size_t index ) const override
   {
-    return (shared_.count(index) > 0);
+    return ( shared_.count( index ) > 0 );
   }
 
   ///
-  inline bool is_ghost(uint index) const override
+  inline bool is_ghost( size_t index ) const override
   {
-    return (ghosts_.count(index) > 0);
+    return ( ghosts_.count( index ) > 0 );
   }
 
   ///
   inline std::string description() const override
   {
-    return std::string("Dof numbering for CG1 scalar");
+    return std::string( "Dof numbering for CG1 scalar" );
   }
 
 private:
-
   ///
-  _set<uint> shared_;
-  _set<uint> ghosts_;
-
+  _set< size_t > shared_;
+  _set< size_t > ghosts_;
 };
 
 }
