+++ include/dolfin/function/GenericFunction.h
@@ -25,7 +25,6 @@ class GenericFunction : public Coefficient, public Variable
 {
 
 public:
-
   /// Constructor
   GenericFunction() = default;
 
@@ -35,35 +34,39 @@ public:
   //--- UFC INTERFACE ---------------------------------------------------------
 
   /// Evaluate function at given point in cell
-  void evaluate(real* values, const real* coordinates,
-                        const ufc::cell& cell) const override = 0;
+  void evaluate( real *            values,
+                 const real *      coordinates,
+                 const ufc::cell & cell ) const override = 0;
 
   //--- INTERFACE -------------------------------------------------------------
 
   /// Evaluate function at given point
-  void eval(real* values, const real* x) const override = 0;
+  void eval( real * values, const real * x ) const override = 0;
 
   /// Return the rank of the value space
-  auto rank() const -> uint override = 0;
+  auto rank() const -> size_t override = 0;
 
   /// Return the dimension of the value space for axis i
-  auto dim(uint i) const -> uint override = 0;
+  auto dim( size_t i ) const -> size_t override = 0;
 
   // Return the value size
-  auto value_size() const -> uint override = 0;
+  auto value_size() const -> size_t override = 0;
 
   /// Interpolate function to vertices of mesh
-  virtual void interpolate_vertex_values(real* values) const = 0;
+  virtual void interpolate_vertex_values( real * values ) const = 0;
 
   /// Interpolate function to finite element space on cell
-  void interpolate(real* coefficients, const ufc::cell& cell,
-                           const ufc::finite_element& finite_element,
-                           const Cell& dolfin_cell) const override = 0;
+  void interpolate( real *                      coefficients,
+                    const ufc::cell &           cell,
+                    const ufc::finite_element & finite_element,
+                    const Cell & dolfin_cell ) const override = 0;
 
   /// Interpolate function to finite element space on facet
-  void interpolate(real* coefficients, const ufc::cell& cell,
-                           const ufc::finite_element& finite_element,
-                           const Cell& dolfin_cell, uint facet) const override = 0;
+  void interpolate( real *                      coefficients,
+                    const ufc::cell &           cell,
+                    const ufc::finite_element & finite_element,
+                    const Cell &                dolfin_cell,
+                    size_t                      facet ) const override = 0;
 
   /// Synchronize
   void sync() override = 0;
@@ -74,24 +77,22 @@ public:
   //---------------------------------------------------------------------------
 
   /// Return the mesh
-  virtual auto mesh() const -> Mesh& = 0;
+  virtual auto mesh() const -> Mesh & = 0;
 
   //---------------------------------------------------------------------------
 
   /// Delegate time dependency
-  auto operator()(Time const& t) -> GenericFunction&
+  auto operator()( Time const & t ) -> GenericFunction &
   {
-    this->sync(t);
+    this->sync( t );
     return *this;
   }
 
   //---------------------------------------------------------------------------
 
 private:
-
   /// Time dependency hook
-  void sync(Time const& t) override = 0;
-
+  void sync( Time const & t ) override = 0;
 };
 
 } /* namespace dolfin */
