+++ include/dolfin/la/GenericSparsityPattern.h
@@ -25,13 +25,13 @@ public:
   virtual ~GenericSparsityPattern() = default;
 
   /// Initialize sparsity pattern for a generic tensor
-  virtual void init(uint rank, uint const * dim, uint const * range) = 0;
+  virtual void init(size_t rank, size_t const * dim, size_t const * range) = 0;
 
   /// Insert non-zero entry
-  virtual void insert(uint const * num_rows, uint const * const * rows) = 0;
+  virtual void insert(size_t const * num_rows, size_t const * const * rows) = 0;
 
   /// Return local size
-  virtual auto size(uint i) const -> uint = 0;
+  virtual auto size(size_t i) const -> size_t = 0;
 
   /// Finalize sparsity pattern (needed by most parallel la backends)
   virtual void apply() = 0;
@@ -46,15 +46,15 @@ public:
   virtual auto is_distributed() const -> bool = 0;
 
   /// Return array with number of non-zeroes per row
-  virtual void numNonZeroPerRow(uint nzrow[]) const = 0;
+  virtual void numNonZeroPerRow(size_t nzrow[]) const = 0;
 
   /// Return array with number of non-zeroes per row split between
   /// diagonal and off-diagonal for process rank
-  virtual void numNonZeroPerRow(uint p_rank, uint d_nzrow[],
-                                uint o_nzrow[]) const = 0;
+  virtual void numNonZeroPerRow(size_t p_rank, size_t d_nzrow[],
+                                size_t o_nzrow[]) const = 0;
 
   /// Return total number of non-zeroes
-  virtual auto numNonZero() const -> uint = 0;
+  virtual auto numNonZero() const -> size_t = 0;
 
   /// Display sparsity pattern
   virtual void disp() const = 0;
